pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.1",
    "numpy>=2.2.3",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "scikit-image>=0.25.2",
    "scipy>=1.15.2",
    "streamlit>=1.43.2",
    "tifffile>=2025.2.18",
]
